add a kinda hacky way for pci bus attachements to enable iommus.
This callback is called by the 0 bus after it has enumerated the base
bus.  Anything that needs resources allocating should have been hit by
that enumeration.

If we see cases where that dosen't work we'll have to do a full bus
heirarchy walk, but that would be rather more complex.